Add new APIs and fix StartHome parameter type
Changes:
- Fix StartHome parameter type from short to int (matches WMX 1.5)
- Add ExtMotion2 StartJerkCoordinatedPos(CoordinatedPosBlockExt2*) block API
- Add EC_Network_Api ecRawWrite/ecRawRead for slave register access
  - Maps to WMX3 RegisterWrite/RegisterRead
  - Export ordinals @270, @271
- Add CoordinatedPosBlockExt2 structure definition with include guards

+++ broker/HomeBroker.cpp
@@ -18,7 +18,7 @@ namespace home {
     {
     }
 
-    WMXAPIFUNC Home::StartHome(short axis)
+    WMXAPIFUNC Home::StartHome(int axis)
     {
         if (!wmxlib) return -1;
 
@@ -28,7 +28,7 @@ namespace home {
         return coreMotion->home->StartHome(axis);
     }
 
-    WMXAPIFUNC Home::StartHome(short firstAxis, short lastAxis)
+    WMXAPIFUNC Home::StartHome(int firstAxis, int lastAxis)
     {
         if (!wmxlib) return -1;
